mgr: only copy keys with dedicated mgr
When collocating mon and mgr, the mgr container will attempt to create
its own key since it has the admin key at its disposal. Also at this
point there is nothing to fetch since the key is not created by the
mons, as mentionned above the mgr creates the key on its own.

This means a wider change where we still deploy mon one by one with
serial:1 but we need to actually stop waiting for the 3 mons to be there
to achieve quorum.

Basically, we must be able to:

* deploy 1 mons with its own quorum
* deploy a second, which will join the existing cluster

Signed-off-by: Sébastien Han <seb@redhat.com>